rewrite all to new get
